+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/util/FieldConverters.java
@@ -30,15 +30,6 @@ public interface FieldConverters {
     SketchAggregation asInnerSketch(SketchAggregation candidate);
 
     /**
-     * Get the candidate SketchAggregation as a SketchEstimatePostAggregation.
-     *
-     * @param candidate  SketchAggregation to "convert"
-     *
-     * @return The SketchAggregation as a SketchEstimatePostAggregation
-     */
-    FuzzySetPostAggregation asSketchEstimate(SketchAggregation candidate);
-
-    /**
      * Get the candidate MetricField as a PostAggregation.
      *
      * @param field  Metric field which can be aggregate or post aggregate