+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/util/SketchFieldConverter.java
@@ -35,7 +35,13 @@ public class SketchFieldConverter implements FieldConverters {
         return new SketchMergeAggregation(candidate.getName(), candidate.getFieldName(), candidate.getSize());
     }
 
-    @Override
+    /**
+     * Get the candidate PostAggregation as a SketchEstimatePostAggregation.
+     *
+     * @param candidate  PostAggregation to "convert"
+     *
+     * @return The PostAggregation as a SketchEstimatePostAggregation
+     */
     public SketchEstimatePostAggregation asSketchEstimate(SketchAggregation candidate) {
         String name = candidate.getName() + "_estimate";
         PostAggregation field = new FieldAccessorPostAggregation(candidate);